feat: add matrix algorithm task
